+++ lib/logstash/inputs/snmptrap.rb
@@ -46,7 +46,7 @@ class LogStash::Inputs::Snmptrap < LogStash::Inputs::Base
   config :supported_transports, :validate => %w[tcp udp], :default => %w[udp], :required => true, :list => true
 
   # The supported SNMP versions to listen on
-  config :supported_versions, :validate => %w[1 2c], default: %w[1 2c], :required => true, :list => true
+  config :supported_versions, :validate => %w[1 2c 3], default: %w[1 2c], :required => true, :list => true
 
   # SNMP Community String to listen for.
   config :community, :validate => :array, :default => 'public'
@@ -64,7 +64,7 @@ class LogStash::Inputs::Snmptrap < LogStash::Inputs::Base
     ::File.join(MIB_BASE_PATH, 'ietf', 'UDP-MIB.dic')
   ].map { |path| ::File.expand_path(path) }
 
-  def initialize(params={})
+  def initialize(params = {})
     super(params)
 
     setup_deprecated_params!
@@ -72,17 +72,18 @@ class LogStash::Inputs::Snmptrap < LogStash::Inputs::Base
   end
 
   def register
+    validate_config!
     mib_manager = build_mib_manager!
 
     if !@mib_paths && !@use_provided_mibs
-      logger.info('Loading default MIB files', :path => MIB_DEFAULT_PATHS)
+      logger.info("Using default MIB paths #{MIB_DEFAULT_PATHS}")
       MIB_DEFAULT_PATHS.each do |path|
         mib_manager.add(path)
       end
     end
 
     @client = build_client!(mib_manager)
-  end # def register
+  end
 
   def run(output_queue)
     begin
@@ -92,8 +93,8 @@ class LogStash::Inputs::Snmptrap < LogStash::Inputs::Base
       @logger.warn('SNMP Trap listener died', format_log_data(e))
       Stud.stoppable_sleep(5) { stop? }
       retry if !stop?
-    end # begin
-  end # def run
+    end
+  end
 
   def stop
     begin
@@ -105,13 +106,19 @@ class LogStash::Inputs::Snmptrap < LogStash::Inputs::Base
 
   private
 
+  def validate_config!
+    if !@security_name.nil? && !@supported_versions.include?('3')
+      raise(LogStash::ConfigurationError, "Using a `security_name` requires `supported_versions` to include version '3': #{@supported_versions}")
+    end
+  end
+
   def build_client!(mib_manager)
-    org.logstash.snmp.SnmpClient
-        .builder(mib_manager, @supported_transports.to_set, @port)
-        .setSupportedVersions(@supported_versions.to_set)
-        .setThreadPoolName('SnmpTrapWorker')
-        .setMapOidVariableValues(@oid_map_field_values)
-        .build
+    client_builder = org.logstash.snmp.SnmpClient
+                     .builder(mib_manager, @supported_transports.to_set, @port)
+                     .setSupportedVersions(@supported_versions.to_set)
+                     .setThreadPoolName('SnmpTrapWorker')
+
+    build_snmp_client!(client_builder, validate_usm_user: @supported_versions.include?('3'))
   end
 
   def consume_trap_message(output_queue, trap_message)
@@ -168,4 +175,4 @@ class LogStash::Inputs::Snmptrap < LogStash::Inputs::Base
     data.merge!(extra_data)
     data
   end
-end # class LogStash::Inputs::Snmptrap
+end