+++ lib/logstash/plugin_mixins/snmp/common.rb
@@ -1,6 +1,8 @@
 module LogStash
   module PluginMixins
     module Snmp
+      # This module provides common configurations and functions to all SNMP plugins
+      #
       module Common
         require 'logstash-integration-snmp_jars'
 
@@ -66,22 +68,30 @@ module LogStash
           # If this setting is omitted, data gets stored at the root (top level) of the event.
           # The target is only relevant while decoding data into a new event.
           base.config :target, :validate => :field_reference
-        end
 
-        def self.default_oid_mapping_format(base)
-          if snmptrap_plugin?(base)
-            OID_MAPPING_FORMAT_RUBY_SNMP
-          else
-            OID_MAPPING_FORMAT_DEFAULT
-          end
-        end
+          # SNMPv3 Credentials
+          #
+          # A single user can be configured and will be used for all defined SNMPv3 requests.
+          # Multiple snmp input declarations will be needed if multiple SNMPv3 users are required.
+          # If not using SNMPv3 simply leave options empty.
 
-        def self.default_use_provided_mibs(base)
-          !snmptrap_plugin?(base)
-        end
+          # The SNMPv3 security name or user name
+          base.config :security_name, :validate => :string
 
-        def self.default_oid_map_field_values(base)
-          snmptrap_plugin?(base)
+          # The SNMPv3 authentication protocol or type
+          base.config :auth_protocol, :validate => %w[md5 sha sha2 hmac128sha224 hmac192sha256 hmac256sha384 hmac384sha512]
+
+          # The SNMPv3 authentication passphrase or password
+          base.config :auth_pass, :validate => :password
+
+          # The SNMPv3 privacy/encryption protocol
+          base.config :priv_protocol, :validate => %w[des 3des aes aes128 aes192 aes256]
+
+          # The SNMPv3 encryption password
+          base.config :priv_pass, :validate => :password
+
+          # The SNMPv3 security level can be Authentication, No Privacy; Authentication, Privacy; or no Authentication, no Privacy
+          base.config :security_level, :validate => %w[noAuthNoPriv authNoPriv authPriv]
         end
 
         def build_mib_manager!
@@ -102,6 +112,17 @@ module LogStash
           end
 
           mib_manager
+        end
+
+        def build_snmp_client!(client_builder, validate_usm_user: false)
+          validate_usm_user! if validate_usm_user
+
+          unless @security_name.nil?
+            client_builder.addUsmUser(@security_name, @auth_protocol, @auth_pass&.value, @priv_protocol, @priv_pass&.value)
+          end
+
+          client_builder.setMapOidVariableValues(@oid_map_field_values)
+          client_builder.build
         end
 
         private
@@ -130,9 +151,74 @@ module LogStash
           end
         end
 
+        def validate_usm_user!
+          errors = []
+
+          errors << '`security_name` is required when SNMP v3 is enabled' if @security_name.nil?
+          validate_auth_protocol(errors)
+          validate_priv_protocol(errors)
+
+          raise(LogStash::ConfigurationError, errors.join(', ')) if errors.any?
+        end
+
+        def validate_auth_protocol(errors)
+          validate_protocol_config(errors, 'auth_protocol', 'auth_pass')
+
+          if @auth_protocol.nil? && (@security_level == 'authNoPriv' || @security_level == 'authPriv')
+            errors << "Using `security_level` set to `#{@security_level}` requires the configuration of `auth_protocol`"
+          end
+        end
+
+        def validate_priv_protocol(errors)
+          validate_protocol_config(errors, 'priv_protocol', 'priv_pass')
+
+          if @priv_protocol.nil? && @security_level == 'authPriv'
+            errors << "Using `security_level` set to `#{@security_level}` requires the configuration of `priv_protocol`"
+          end
+        end
+
+        def validate_protocol_config(errors, protocol_config_name, pass_config_name)
+          protocol_config_value = instance_variable_get("@#{protocol_config_name}".intern)
+          pass_config_value = instance_variable_get("@#{pass_config_name}".intern)
+
+          if protocol_config_value && !pass_config_value
+            errors << "Using `#{protocol_config_name}` requires the `#{pass_config_name}`"
+          elsif pass_config_value && !protocol_config_value
+            errors << "`#{protocol_config_name}` is required when using `#{pass_config_name}`"
+          end
+
+          if pass_config_value&.value && pass_config_value.value.length < 8
+            errors << "`#{pass_config_name}` passphrase must be at least 8 bytes long"
+          end
+        end
+
         def self.snmptrap_plugin?(base)
           !defined?(LogStash::Inputs::Snmptrap).nil? && base == LogStash::Inputs::Snmptrap
         end
+
+        private_class_method :snmptrap_plugin?
+
+        def self.default_oid_mapping_format(base)
+          if snmptrap_plugin?(base)
+            OID_MAPPING_FORMAT_RUBY_SNMP
+          else
+            OID_MAPPING_FORMAT_DEFAULT
+          end
+        end
+
+        private_class_method :default_oid_mapping_format
+
+        def self.default_use_provided_mibs(base)
+          !snmptrap_plugin?(base)
+        end
+
+        private_class_method :default_use_provided_mibs
+
+        def self.default_oid_map_field_values(base)
+          snmptrap_plugin?(base)
+        end
+
+        private_class_method :default_oid_map_field_values
       end
     end
   end